Move MaasTheme color palette and corner radii implementations to separate files.
